move interpolation function from reconst to core
